OpenGL_libpq_Draw - now the index is uint and not short anymore

+++ src/OpenGL_libpq_Draw/ogl.h
@@ -47,7 +47,7 @@ private:
   std::vector<glm::vec2>        uvs;
   std::vector<glm::vec3>        normals;
 
-  std::vector<unsigned short>   indices;
+  std::vector<unsigned int>     indices;
   std::vector<glm::vec3>        indexed_vertices;
   std::vector<glm::vec2>        indexed_uvs;
   std::vector<glm::vec3>        indexed_normals;
